requirements.txt
flask==2.3.3
google-api-python-client==2.100.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
requests==2.31.0
pytz==2023.3 